Move new files to `state` package
