+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/flowcontroller/FlowControllerViewModel.kt
@@ -3,6 +3,7 @@ package com.stripe.android.paymentsheet.flowcontroller
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import com.stripe.android.paymentsheet.model.PaymentSelection
+import com.stripe.android.paymentsheet.state.PaymentSheetState
 
 internal class FlowControllerViewModel(
     private val handle: SavedStateHandle
