+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/injection/FlowControllerModule.kt
@@ -9,9 +9,9 @@ import com.stripe.android.paymentsheet.DefaultPrefsRepository
 import com.stripe.android.paymentsheet.PaymentSheet
 import com.stripe.android.paymentsheet.PrefsRepository
 import com.stripe.android.paymentsheet.analytics.EventReporter
-import com.stripe.android.paymentsheet.flowcontroller.DefaultPaymentSheetLoader
+import com.stripe.android.paymentsheet.state.DefaultPaymentSheetLoader
 import com.stripe.android.paymentsheet.flowcontroller.FlowControllerViewModel
-import com.stripe.android.paymentsheet.flowcontroller.PaymentSheetLoader
+import com.stripe.android.paymentsheet.state.PaymentSheetLoader
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
