+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/state/DefaultPaymentSheetLoaderTest.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.paymentsheet.flowcontroller
+package com.stripe.android.paymentsheet.state
 
 import android.app.Application
 import androidx.test.core.app.ApplicationProvider
