+++ paymentsheet/src/test/java/com/stripe/android/utils/FakePaymentSheetLoader.kt
@@ -4,8 +4,8 @@ import com.stripe.android.model.PaymentIntentFixtures
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.StripeIntent
 import com.stripe.android.paymentsheet.PaymentSheet
-import com.stripe.android.paymentsheet.flowcontroller.PaymentSheetLoader
-import com.stripe.android.paymentsheet.flowcontroller.PaymentSheetState
+import com.stripe.android.paymentsheet.state.PaymentSheetLoader
+import com.stripe.android.paymentsheet.state.PaymentSheetState
 import com.stripe.android.paymentsheet.model.ClientSecret
 import com.stripe.android.paymentsheet.model.SavedSelection
 import kotlinx.coroutines.delay
